Move UIViewKitDevelopmentViews to testTarget and update Carthage project
