+++ Sources/UIViewDSL/File.swift
@@ -1,0 +1,9 @@
+//
+//  File.swift
+//  
+//
+//  Created by MaxAir on 12/02/2024.
+//
+
+import Foundation
+public let toto = "toto test"
