+++ Sources/UIViewDSL/UIViewDSL.h
@@ -1,0 +1,18 @@
+//
+//  UIViewDSL.h
+//  UIViewDSL
+//
+//  Created by MaxAir on 12/02/2024.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for UIViewDSL.
+FOUNDATION_EXPORT double UIViewDSLVersionNumber;
+
+//! Project version string for UIViewDSL.
+FOUNDATION_EXPORT const unsigned char UIViewDSLVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <UIViewDSL/PublicHeader.h>
+
+
