+++ Sources/UIViewKit/Debug/IBDebug.swift
@@ -6,6 +6,14 @@
 //
 
 import UIKit
+import UIViewDSL
+
+extension UIViewDSL where Self: UIView {
+    public func ibDebug() -> Self {
+        self.layer.borderWidth = 1
+        return self
+    }
+}
 
 public class IBDebug {
