+++ Sources/UIViewKit/UIKitExtensions/UIStackView+Extensions.swift
@@ -1,0 +1,25 @@
+//
+//  UIStackView+Extensions.swift
+//  UIViewKit
+//
+//  Created by Blazej SLEBODA on 29/09/2023.
+//
+
+import UIKit
+
+extension UIStackView {
+    
+    public convenience init(axis: NSLayoutConstraint.Axis, spacing: CGFloat? = nil, alignment: UIStackView.Alignment? = nil, distribution: UIStackView.Distribution? = nil) {
+        self.init()
+        self.axis = axis
+        if let spacing {
+            self.spacing = spacing
+        }
+        if let alignment {
+            self.alignment = alignment
+        }
+        if let distribution {
+            self.distribution = distribution
+        }
+    }
+}
