+++ Tests/UIViewKitTests/UIViewKitDevelopmentViews/HelloWordView.swift
@@ -7,7 +7,6 @@
 
 import UIKit
 import UIViewKit
-import UIViewDSL
 
 public class HelloWordView: UIView {
 
@@ -17,7 +16,7 @@ public class HelloWordView: UIView {
         super.init(frame: frame)
         self {
             UILabel().ibOutlet(&label).ibAttributes {
-                IBConstraints.createConstraints(from: $0, to: self, guide: .view, anchors: .centerX, .centerY)
+                IBConstraints.create(from: $0, to: self, guide: .view, anchors: .centerX, .centerY)
                 $0.text = "Hello, world!"
             }
         }.ibAttributes {
@@ -27,12 +26,3 @@ public class HelloWordView: UIView {
 
     required init?(coder: NSCoder) { fatalError() }
 }
-
-#if DEBUG
-
-@available(iOS 17.0, *)
-#Preview("Default") {
-    HelloWordView()
-}
-
-#endif
