+++ Tests/UIViewKitTests/UIViewKitDevelopmentViews/UserTitleTimeDescriptionView/UserTitleTimeDescriptionViewWithIBSubviewsWithSuperviewView.swift
@@ -1,5 +1,5 @@
 //
-//  UserTitleTimeDescriptionViewWithIBSubviews.swift
+//  UserTitleTimeDescriptionViewWithIBSubviewsWithSuperviewView.swift
 //  UIViewKit
 //
 //  Created by Blazej SLEBODA on 18/09/2023.
@@ -7,9 +7,8 @@
 
 import UIKit
 import UIViewKit
-import SwiftUI
 
-public class UserTitleTimeDescriptionViewWithIBSubviews: UIView {
+public class UserTitleTimeDescriptionViewWithIBSubviewsWithSuperviewView: UIView {
 
     var imageViewUser: UIImageView!
     var labelTime: UILabel!
@@ -19,19 +18,19 @@ public class UserTitleTimeDescriptionViewWithIBSubviews: UIView {
     override init(frame: CGRect) {
         super.init(frame: frame)
 
-        self.ibSubviews {
-            UIStackView(axis: .horizontal, spacing: 12, alignment: .top).ibSubviews {
+        self.ibSubviews { _ in
+            UIStackView(axis: .horizontal, spacing: 12, alignment: .top).ibSubviews { _ in
                 UIImageView().ibOutlet(&imageViewUser).ibAttributes {
                     $0.widthAnchor.constraint(equalToConstant: 60)
                     $0.heightAnchor.constraint(equalToConstant: 60)
                     $0.setContentHuggingPriority(.defaultHigh + 1, for: .horizontal)
                 }
-                UIStackView(axis: .vertical).ibSubviews {
-                    UIStackView(axis: .horizontal, spacing: 8, alignment: .center).ibSubviews {
+                UIStackView(axis: .vertical).ibSubviews { _ in
+                    UIStackView(axis: .horizontal, spacing: 8, alignment: .center).ibSubviews { _ in
                         UILabel().ibOutlet(&labelTitle).ibAttributes {
                             $0.font = .systemFont(ofSize: 20, weight: .semibold)
                         }
-                        UIStackView(axis: .horizontal, spacing: 8).ibSubviews {
+                        UIStackView(axis: .horizontal, spacing: 8).ibSubviews { _ in
                             UILabel().ibOutlet(&labelTime).ibAttributes {
                                 $0.textColor = .systemGray
                                 $0.font = .systemFont(ofSize: 15, weight: .semibold)
@@ -62,30 +61,10 @@ public class UserTitleTimeDescriptionViewWithIBSubviews: UIView {
                     }
                 }
             }.ibAttributes {
-                IBConstraints.createConstraints(from: $0, to: self, guide: .view, anchors: .top, .left(12), .right(-12), .bottom)
+                IBConstraints.create(from: $0, to: self, guide: .view, anchors: .top, .left(12), .right(-12), .bottom)
             }
         }
     }
 
     required init?(coder: NSCoder) { fatalError() }
 }
-
-#if DEBUG
-
-import SwiftUI
-
-struct UserTitleTimeDescriptionViewWithIBSubviewsPreviews: PreviewProvider {
-    static var previews: some View {
-        IBPreview.FullScreenView(
-            UserTitleTimeDescriptionViewWithIBSubviews().ibAttributes {
-                $0.imageViewUser.image = .init(systemName: "person.circle")
-                $0.labelTitle.text = "Amanda Clarke"
-                $0.labelTime.text = "1:08 PM"
-                $0.labelDescription.text = "Hey, how are you doing today? I have got a question about our trip to Mexi..."
-                IBDebug.showFrames(of: $0, includeGivenView: true, includeUIKitPrivateViews: false)
-            }
-        )
-    }
-}
-
-#endif
