+++ Tests/UIViewKitTests/UIViewKitDevelopmentViews/ViewSubviewHasConstraintsToItsSuperviewN2.swift
@@ -7,7 +7,6 @@
 
 import UIKit
 import UIViewKit
-import UIViewDSL
 
 class TestView: UIView {
 
@@ -20,7 +19,7 @@ class TestView: UIView {
             view2 {
                 view3.ibAttributes {
                     $0.tag = 3
-                    IBConstraints.createConstraints(from: $0, to: self, guide: .view, anchors: .all)
+                    IBConstraints.create(from: $0, to: self, guide: .view, anchors: .all)
                     $0.backgroundColor = .green
                     $0.alpha = 0.3
                 }
@@ -40,16 +39,3 @@ class TestView: UIView {
         fatalError()
     }
 }
-
-#if DEBUG
-
-import SwiftUI
-
-@available(iOS 14.0, *)
-struct TestViewPreviews: PreviewProvider {
-    static var previews: some View {
-        IBPreview.FullScreenView(TestView()).ignoresSafeArea()
-    }
-}
-
-#endif
