+++ Tests/UIViewKitTests/UserTitleTimeDescriptionViewTests/UserTitleTimeDescriptionViewTests.swift
@@ -7,7 +7,6 @@
 
 import XCTest
 @testable import UIViewKit
-@testable import UIViewKitDevelopmentViews
 
 class UserTitleTimeDescriptionViewTests: XCTestCase {
